requirements.txt
asgiref==3.5.2
dj-database-url==0.5.0
Django==4.0.5
django-credit-cards==0.4.1
gunicorn==20.1.0
Pillow==9.1.1
psycopg2==2.9.3
sqlparse==0.4.2
tzdata==2022.1
whitenoise==6.2.0
